Remove duplicated methd by moving it into an extension
