+++ Sources/Request Strategies/User/UserProfileRequestStrategyTests.swift
@@ -438,18 +438,6 @@ class UserProfileRequestStrategyTests: MessagingTestBase {
         return response
     }
 
-    func responseFailure(code: Int, label: Payload.ResponseFailure.Label, message: String = "") -> ZMTransportResponse {
-        let responseFailure = Payload.ResponseFailure(code: code, label: label, message: message)
-        let payloadData = responseFailure.payloadData()!
-        let payloadString = String(bytes: payloadData, encoding: .utf8)!
-        let response = ZMTransportResponse(payload: payloadString as ZMTransportData,
-                                           httpStatus: code,
-                                           transportSessionError: nil)
-
-        return response
-
-    }
-
     func userProfile(for uuid: UUID, domain: String?) -> Payload.UserProfile {
         return Payload.UserProfile(id: uuid,
                                    qualifiedID: nil,
